% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/parallel_4.R
\name{parallel_4}
\alias{parallel_4}
\title{A simple function for running Stan in parallel}
\usage{
parallel_4(fit, data, iter = 2000, pars = c("alpha", "beta", "a",
  "log_lik"), cores = 4)
}
\description{
A simple function for running Stan in parallel
}

