% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/stan_predict_speeches.R
\name{predict_prob}
\alias{predict_prob}
\title{Predict probabilities for multiple scenarios}
\usage{
predict_prob(stanfit, data, fitted_values, a_num, betas,
  model_pars = c("beta", "alpha", "a"))
}
\arguments{
\item{stanfit}{a \code{stanfit} object.}

\item{data}{the data used to create the \code{stanfit} object.}

\item{fitted_values}{a matrix of fitted values for the \code{betas}.
Each row should be a different scenario.}

\item{a_num}{integer of the \code{a} speaker level to fit for.}

\item{model_pars}{character vector of model parameters to use in in the
predicted probabilities. Note: should include parameters beginning with
"\code{beta}" that correspond to the fitted values,}
}
\description{
Predict probabilities for multiple scenarios
}

