% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/utils.R
\name{stan_lister}
\alias{stan_lister}
\title{Create data lists for Stan from a base list and model covariates}
\usage{
stan_lister(base, df, vars)
}
\arguments{
\item{base}{a list with the shared elements.}

\item{df}{a data frame where the \code{vars} can be found.}

\item{vars}{a character vector of variable names to include in the model.}
}
\description{
Create data lists for Stan from a base list and model covariates
}

