% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/stan_speeches_table.R
\name{stan_speeches_param_est}
\alias{stan_speeches_param_est}
\title{Create data frame of results from stanfit speeches for tables using xtable}
\usage{
stan_speeches_param_est(stanfit, model_pars = c("beta", "alpha"), pars_labels,
  col_labels, obs)
}
\arguments{
\item{stanfit}{list of stanfit objects. Names will become column labels
unless \code{col_labels} is specified.}

\item{model_pars}{character vector of parameters from the \code{stanfit}
object to include in the table.}

\item{pars_labels}{required vector of parameter labels to include in the
table. Must be a list with parameter labels for each element in
\code{stanfit} and in the same order. Identical variables must have matching
labels.}

\item{col_labels}{optional vector of column labels. If not specified then
names of fitted stan models in \code{stanfit} are used.}

\item{obs}{intiger number of observations in the model. Note: currently crude
implementation.}
}
\description{
Create data frame of results from stanfit speeches for tables using xtable
}

